+++ components/notifier/tests/strategies/test_changed_status.py
@@ -5,7 +5,7 @@ import json
 import pathlib
 import unittest
 
-from strategies.changed_status import get_notable_metrics_from_json, has_new_status
+from strategies.changed_status import get_notable_metrics_from_json, status_changed
 
 
 class StrategiesTestCase(unittest.TestCase):
@@ -22,8 +22,8 @@ class StrategiesTestCase(unittest.TestCase):
     def setUp(self):
         """Set variables for the other testcases."""
         self.most_recent_measurement_seen = datetime.datetime.min.isoformat()
-        self.first_timestamp = "2019-01-01T00:23:59+59:00"
-        self.second_timestamp = "2020-01-01T00:23:59+59:00"
+        self.old_timestamp = "2019-01-01T00:23:59+59:00"
+        self.new_timestamp = "2020-01-01T00:23:59+59:00"
         self.report_url = "https://report1"
         self.white_metric_status = "unknown"
         self.red_metric_status = "red (target not met)"
@@ -51,7 +51,7 @@ class StrategiesTestCase(unittest.TestCase):
         """Test that there is nothing to notify when there are no red metrics."""
         count = dict(status="target_met", value="0")
         green_metric = self.metric(
-            recent_measurements=[dict(start=self.first_timestamp, end=self.second_timestamp, count=count)]
+            recent_measurements=[dict(start=self.old_timestamp, end=self.new_timestamp, count=count)]
         )
         subject1 = dict(metrics=dict(metric1=green_metric))
         report1 = dict(report_uuid="report1", title="report_title", subjects=dict(subject1=subject1))
@@ -65,7 +65,7 @@ class StrategiesTestCase(unittest.TestCase):
         count = dict(status="target_not_met", value="10")
         red_metric = self.metric(
             status="target_not_met",
-            recent_measurements=[dict(start=self.first_timestamp, end=self.second_timestamp, count=count)],
+            recent_measurements=[dict(start=self.old_timestamp, end=self.new_timestamp, count=count)],
         )
         subject1 = dict(metrics=dict(metric1=red_metric))
         report1 = dict(report_uuid="report1", title="Title", subjects=dict(subject1=subject1))
@@ -92,8 +92,8 @@ class StrategiesTestCase(unittest.TestCase):
         red_metric = self.metric(
             status="target_not_met",
             recent_measurements=[
-                dict(start=self.first_timestamp, end=self.second_timestamp, count=old_count),
-                dict(start=self.second_timestamp, end=self.second_timestamp, count=new_count),
+                dict(start=self.old_timestamp, end=self.new_timestamp, count=old_count),
+                dict(start=self.new_timestamp, end=self.new_timestamp, count=new_count),
             ],
         )
         subject1 = dict(metrics=dict(metric1=red_metric))
@@ -128,8 +128,8 @@ class StrategiesTestCase(unittest.TestCase):
             scale="percentage",
             status="target_not_met",
             recent_measurements=[
-                dict(start=self.first_timestamp, end=self.second_timestamp, percentage=old_percentage),
-                dict(start=self.second_timestamp, end=self.second_timestamp, percentage=new_percentage),
+                dict(start=self.old_timestamp, end=self.new_timestamp, percentage=old_percentage),
+                dict(start=self.new_timestamp, end=self.new_timestamp, percentage=new_percentage),
             ],
         )
         subject1 = dict(metrics=dict(metric1=red_metric))
@@ -155,50 +155,45 @@ class StrategiesTestCase(unittest.TestCase):
             get_notable_metrics_from_json(self.data_model, reports_json, self.most_recent_measurement_seen)[0].metrics,
         )
 
-    def test_new_white_metric(self):
+    def test_recently_changed_metric_status(self):
         """Test that a metric that turns white is added."""
         metric = self.metric(
             status=self.white_metric_status,
             recent_measurements=[
-                dict(start=self.second_timestamp, count=dict(status="target_met")),
-                dict(start=self.first_timestamp, count=dict(status=self.white_metric_status)),
-            ],
-        )
-        self.assertTrue(has_new_status(metric, self.most_recent_measurement_seen, self.white_metric_status))
-        self.assertFalse(has_new_status(metric, self.most_recent_measurement_seen, self.red_metric_status))
-        self.assertTrue(
-            has_new_status(metric, self.most_recent_measurement_seen, self.white_metric_status, self.red_metric_status)
-        )
+                dict(start=self.new_timestamp, count=dict(status="target_met")),
+                dict(start=self.old_timestamp, count=dict(status=self.white_metric_status))])
+        self.assertTrue(status_changed(metric, self.most_recent_measurement_seen))
 
-    def test_old_white_metric(self):
+    def test_new_measurement_same_status(self):
         """Test that a metric that was already white isn't added."""
         metric = self.metric(
             status=self.white_metric_status,
             recent_measurements=[
-                dict(start=self.second_timestamp, count=dict(status=self.white_metric_status)),
-                dict(start=self.first_timestamp, count=dict(status=self.white_metric_status)),
-            ],
-        )
-        self.assertFalse(
-            has_new_status(metric, self.most_recent_measurement_seen, self.white_metric_status, self.red_metric_status)
-        )
+                dict(start=self.new_timestamp, count=dict(status=self.white_metric_status)),
+                dict(start=self.old_timestamp, count=dict(status=self.white_metric_status))])
+        self.assertFalse(status_changed(metric, self.most_recent_measurement_seen))
 
-    def test_only_one_measurement(self):
+    def test_new_measurement_different_status_outside_time_period(self):
+        """Test that a metric that was already white isn't added."""
+        oldest_timestamp = "2018-01-01T00:23:59+59:00"
+        metric = self.metric(
+            status=self.white_metric_status,
+            recent_measurements=[
+                dict(start=self.old_timestamp, count=dict(status=self.white_metric_status)),
+                dict(start=oldest_timestamp, count=dict(status="target_not_met"))])
+        self.assertFalse(status_changed(metric, self.new_timestamp))
+
+    def test_no_change_due_to_only_one_measurement(self):
         """Test that metrics with only one measurement (and therefore no changes in value) aren't added."""
         metric = self.metric(
             status=self.white_metric_status,
-            recent_measurements=[dict(start=self.first_timestamp, count=dict(status=self.white_metric_status))],
-        )
-        self.assertFalse(
-            has_new_status(metric, self.most_recent_measurement_seen, self.white_metric_status, self.red_metric_status)
-        )
+            recent_measurements=[dict(start=self.old_timestamp, count=dict(status=self.white_metric_status))])
+        self.assertFalse(status_changed(metric, self.most_recent_measurement_seen))
 
-    def test_no_measurements(self):
+    def test_no_change_due_to_no_measurements(self):
         """Test that metrics without measurements (and therefore no changes in value) aren't added."""
         metric = self.metric(status=self.white_metric_status)
-        self.assertFalse(
-            has_new_status(metric, self.most_recent_measurement_seen, self.white_metric_status, self.red_metric_status)
-        )
+        self.assertFalse(status_changed(metric, self.most_recent_measurement_seen))
 
     def test_multiple_reports_with_same_destination(self):
         """Test that the correct metrics are notified when multiple reports notify the same destination."""
@@ -208,8 +203,8 @@ class StrategiesTestCase(unittest.TestCase):
         red_metric1 = self.metric(
             status="target_not_met",
             recent_measurements=[
-                dict(start=self.first_timestamp, end=self.second_timestamp, count=old_count),
-                dict(start=self.second_timestamp, end=self.second_timestamp, count=new_count),
+                dict(start=self.old_timestamp, end=self.new_timestamp, count=old_count),
+                dict(start=self.new_timestamp, end=self.new_timestamp, count=new_count),
             ],
         )
         subject1 = dict(metrics=dict(metric1=red_metric1))
@@ -225,8 +220,8 @@ class StrategiesTestCase(unittest.TestCase):
             name="metric2",
             status="target_met",
             recent_measurements=[
-                dict(start=self.first_timestamp, end=self.second_timestamp, count=old_count),
-                dict(start=self.second_timestamp, end=self.second_timestamp, count=new_count),
+                dict(start=self.old_timestamp, end=self.new_timestamp, count=old_count),
+                dict(start=self.new_timestamp, end=self.new_timestamp, count=new_count),
             ],
         )
         subject2 = dict(metrics=dict(metric1=red_metric2))
@@ -250,9 +245,20 @@ class StrategiesTestCase(unittest.TestCase):
                         metric_type="tests",
                         metric_name="metric1",
                         metric_unit="units",
-                        new_metric_status="red (target not met)",
+                        new_metric_status=self.red_metric_status,
                         new_metric_value="10",
-                        old_metric_status="green (target met)",
+                        old_metric_status=self.green_metric_status,
+                        old_metric_value="5",
+                    )
+                ],
+                [
+                    dict(
+                        metric_type="tests",
+                        metric_name="metric2",
+                        metric_unit="units",
+                        new_metric_status=self.red_metric_status,
+                        new_metric_value="10",
+                        old_metric_status=self.green_metric_status,
                         old_metric_value="5",
                     )
                 ]
@@ -268,8 +274,8 @@ class StrategiesTestCase(unittest.TestCase):
             name="metric1",
             status="target_not_met",
             recent_measurements=[
-                dict(start=self.first_timestamp, end=self.second_timestamp, count=old_count),
-                dict(start=self.second_timestamp, end=self.second_timestamp, count=new_count),
+                dict(start=self.old_timestamp, end=self.new_timestamp, count=old_count),
+                dict(start=self.new_timestamp, end=self.new_timestamp, count=new_count),
             ],
         )
         subject1 = dict(metrics=dict(metric1=red_metric))
@@ -293,8 +299,8 @@ class StrategiesTestCase(unittest.TestCase):
             name="metric1",
             status="target_not_met",
             recent_measurements=[
-                dict(start=self.first_timestamp, end=self.second_timestamp, count=old_count),
-                dict(start=self.second_timestamp, end=self.second_timestamp, count=new_count),
+                dict(start=self.old_timestamp, end=self.new_timestamp, count=old_count),
+                dict(start=self.new_timestamp, end=self.new_timestamp, count=new_count),
             ],
         )
         subject1 = dict(metrics=dict(metric1=red_metric))